move styles to src folder
