+++ src/Components/MainHeader.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import logo from "../Images/Msign.png";
-import "../../public/styles/layout/MainHeader.css";
+import "../styles/layout/MainHeader.css";
 
 export default function Header(props) {
   return (
