+++ src/Components/Posts.jsx
@@ -1,10 +1,10 @@
 import React from "react";
-import "../../public/styles/layout/Posts.css";
+import "../styles/layout/Posts.css";
 
 export default function Posts(props) {
   return (
     <div className="posts">
-      <h2 className="post-header">Posts</h2>
+      <h2 className="post-header"> Posts </h2>{" "}
     </div>
   );
 }
