+++ src/Components/Video.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "../../public/styles/layout/Video.css"
+import "../styles/layout/Video.css";
 import Video1 from "../Videos/Video1.mp4";
 
 export default function Video(): JSX.Element {
