+++ src/index.js
@@ -1,8 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import "../../public/styles/pages/index.css";
-import App from "./App";
-import reportWebVitals from "../reportWebVitals";
+import "./styles/pages/index.css";
+import App from "./Pages/App";
+import reportWebVitals from "./reportWebVitals";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
